Move pentagon network test setups to own file
