+++ test/test_helper_pentagons.h
@@ -1,0 +1,10 @@
+#ifndef TEST_HELPER_PENTAGONS_H
+#define TEST_HELPER_PENTAGONS_H
+
+void set_majority_net_pentagon_2(int *majority, int node_ct);
+void set_majority_net_pentagon_3(int *majority, int node_ct);
+void set_majority_net_pentagon_4(int *majority, int node_ct);
+void set_majority_net_pentagon_5(int *majority, int node_ct);
+void set_majority_net_pentagon_6(int *majority, int node_ct);
+
+#endif
